Refactor solid login to resolve locmax lint error
